Switch over to new bindings
